requirements.txt
streamlit
requests
beautifulsoup4
readability-lxml
certifi
fpdf2
PyPDF2
gTTS
deep-translator
openai
python-dotenv
sentence-transformers
scikit-learn
numpy
pandas
plotly
Pillow
selenium==4.15.2
webdriver-manager==4.0.1
playwright==1.48.0
greenlet==3.1.1
